move style.js to styles/
